\name{lambertw}
\alias{lambertw}
\title{
Lambert W Function
}
\description{
The Lambert W function, given an input z, outputs w such that z = w
exp(w). A numerical algorithm of order 4 is used to find the zero of z -
w exp(w). 
}
\usage{
lambertw(z)
}
\arguments{
  \item{z}{ the argument of the Lambert W function
}
}
\value{Returns the value of the Lambert W function.
}
\references{
Corless, R. M., Gonnet, G. H., Hare, D. E. G., Jeffrey, D. J. and Knuth,
D. E. (1996), On the Lambert W Function, 5, Advances in Computational
Mathematics
}
\author{
Julie Carreau
}
